refactor into controller and routes files
